+++ R/interactive_plotting.R
@@ -1,22 +1,24 @@
 #' Plot a 2D slice from an MRSI fit result object.
 #' @param fit_res \code{fit_result} object.
-#' @param name name of the quantity to plot, eg "TNAA".
+#' @param map array of values to be plotted, defaults to a "TNAA" map.
 #' @param slice slice to plot in the z direction.
 #' @param zlim range of values to plot.
 #' @param interp interpolation factor.
 #' @export
-plot_fit_slice_inter <- function(fit_res, name, slice = 1, zlim = NULL, 
+plot_fit_slice_inter <- function(fit_res, map = NULL, slice = 1, zlim = NULL, 
                                  interp = 1) {
   
-  map <- get_fit_map(fit_res, name) 
-  plot_slice_map_inter(map = map, mrs_data = fit_res, slice = slice, 
+  if (is.null(map)) map <- get_fit_map(fit_res, "TNAA") 
+  
+  plot_slice_map_inter(mrs_data = fit_res, map = map, slice = slice, 
                        interp = interp, zlim = zlim)
 }
 
 #' Plot an interactive slice map from a data array where voxels can be selected
 #' to display a corresponding spectrum.
-#' @param map array of values to be plotted.
 #' @param mrs_data spectral data.
+#' @param map array of values to be plotted, defaults to the integration of the
+#' modulus of the full spectral width.
 #' @param xlim spectral region to plot.
 #' @param slice the slice index to plot.
 #' @param zlim smallest and largest values to be plotted.
@@ -27,7 +29,7 @@ plot_fit_slice_inter <- function(fit_res, name, slice = 1, zlim = NULL,
 #' @param interp map interpolation factor.
 #' @export
 #' @importFrom tkrplot tkrplot
-plot_slice_map_inter <- function(map, mrs_data, xlim = NULL, slice = 1,
+plot_slice_map_inter <- function(mrs_data, map = NULL, xlim = NULL, slice = 1,
                                  zlim = NULL, mask_map = NULL, denom = NULL, 
                                  mask_cutoff = 20, interp = 1) {
   
@@ -40,6 +42,8 @@ plot_slice_map_inter <- function(map, mrs_data, xlim = NULL, slice = 1,
 
   assign("plot_env", new.env(hash = TRUE), envir = baseenv())
   #assign("plot_env", new.env(hash = TRUE), envir = globalenv())
+  
+  if (is.null(map)) map <- int_spec(mrs_data, mode = "mod")
 
   if (class(mrs_data) == "mrs_data") {
     x_scale <- ppm(mrs_data)
@@ -71,7 +75,7 @@ plot_slice_map_inter <- function(map, mrs_data, xlim = NULL, slice = 1,
   plot_env$y <- 1
 
 
-  plot_env$win1 <- tcltk::tktoplevel()
+  plot_env$win1 <- tcltk::tktoplevel(class = "spant_plot")
 
   #plot_env$win1$env$plot <- tkrplot::tkrplot(plot_env$win1, fun = plotTk,
   #                                           hscale = 3.0, vscale = 1.5)